+++ lib/designs/atomic/tokens/src/gen/assets.gen.dart
@@ -47,13 +47,49 @@ class $LibAssetsColorsGen {
 class $LibAssetsIconsGen {
   const $LibAssetsIconsGen();
 
+  /// File path: lib/assets/icons/Radiobuttons_no_selected.svg
+  SvgGenImage get radiobuttonsNoSelected =>
+      const SvgGenImage('lib/assets/icons/Radiobuttons_no_selected.svg');
+
+  /// File path: lib/assets/icons/Radiobuttons_selected.svg
+  SvgGenImage get radiobuttonsSelected =>
+      const SvgGenImage('lib/assets/icons/Radiobuttons_selected.svg');
+
+  /// File path: lib/assets/icons/alert_icon.svg
+  SvgGenImage get alertIcon =>
+      const SvgGenImage('lib/assets/icons/alert_icon.svg');
+
+  /// File path: lib/assets/icons/apple_logo.svg
+  SvgGenImage get appleLogo =>
+      const SvgGenImage('lib/assets/icons/apple_logo.svg');
+
   /// File path: lib/assets/icons/arrow_right.svg
   SvgGenImage get arrowRight =>
       const SvgGenImage('lib/assets/icons/arrow_right.svg');
 
+  /// File path: lib/assets/icons/bell_icon.svg
+  SvgGenImage get bellIcon =>
+      const SvgGenImage('lib/assets/icons/bell_icon.svg');
+
+  /// File path: lib/assets/icons/cultural_asset_coin.svg
+  SvgGenImage get culturalAssetCoin =>
+      const SvgGenImage('lib/assets/icons/cultural_asset_coin.svg');
+
+  /// File path: lib/assets/icons/double_right_arrow.svg
+  SvgGenImage get doubleRightArrow =>
+      const SvgGenImage('lib/assets/icons/double_right_arrow.svg');
+
   /// File path: lib/assets/icons/exclamacion.svg
   SvgGenImage get exclamacion =>
       const SvgGenImage('lib/assets/icons/exclamacion.svg');
+
+  /// File path: lib/assets/icons/left_arrow.svg
+  SvgGenImage get leftArrow =>
+      const SvgGenImage('lib/assets/icons/left_arrow.svg');
+
+  /// File path: lib/assets/icons/plus_icon.svg
+  SvgGenImage get plusIcon =>
+      const SvgGenImage('lib/assets/icons/plus_icon.svg');
 
   /// File path: lib/assets/icons/search_navbar.svg
   SvgGenImage get searchNavbar =>
@@ -61,6 +97,10 @@ class $LibAssetsIconsGen {
 
   /// File path: lib/assets/icons/star.svg
   SvgGenImage get star => const SvgGenImage('lib/assets/icons/star.svg');
+
+  /// File path: lib/assets/icons/subsciption_button_icon.svg
+  SvgGenImage get subsciptionButtonIcon =>
+      const SvgGenImage('lib/assets/icons/subsciption_button_icon.svg');
 
   /// File path: lib/assets/icons/subscription_icon.svg
   SvgGenImage get subscriptionIcon =>
@@ -80,10 +120,20 @@ class $LibAssetsIconsGen {
 
   /// List of all assets
   List<SvgGenImage> get values => [
+    radiobuttonsNoSelected,
+    radiobuttonsSelected,
+    alertIcon,
+    appleLogo,
     arrowRight,
+    bellIcon,
+    culturalAssetCoin,
+    doubleRightArrow,
     exclamacion,
+    leftArrow,
+    plusIcon,
     searchNavbar,
     star,
+    subsciptionButtonIcon,
     subscriptionIcon,
     subscriptionIconWeb,
     tagSubscription,
@@ -94,12 +144,19 @@ class $LibAssetsIconsGen {
 class $LibAssetsMocksGen {
   const $LibAssetsMocksGen();
 
+  /// File path: lib/assets/mocks/payment_summary_info_response.json
+  String get paymentSummaryInfoResponse =>
+      'lib/assets/mocks/payment_summary_info_response.json';
+
   /// File path: lib/assets/mocks/subscription_plan_by_user.json
   String get subscriptionPlanByUser =>
       'lib/assets/mocks/subscription_plan_by_user.json';
 
   /// List of all assets
-  List<String> get values => [subscriptionPlanByUser];
+  List<String> get values => [
+    paymentSummaryInfoResponse,
+    subscriptionPlanByUser,
+  ];
 }
 
 class StoycoAssets {